refactor(showcase): single-source shared catalog flatten with manifest validation
One shared catalog-flatten authority (typed throws, not process.exit); the server
re-flatten validates manifest structure at parity with the codegen path.
